@@ -8,7 +8,7 @@ readme = "README.md"
 repository = "https://www.github.com/langchain-ai/langgraph"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
+python = ">=3.9.0,<4.0"
 langchain-core = "^0.1.8"
 
 
